Move LoadFont JavaScript to head for availability
This commit moves the LoadFont JavaScript to the head of the HTML document so
that the load_font function is available when a page wants to use it.
Previously, the LoadFont JavaScript was being loaded in the end of the HTML
document, which meant that it was not always available when needed.